bugs fixed (watchdogs of DPT implementations)

+++ iotsys-common/src/at/ac/tuwien/auto/iotsys/commons/obix/objects/general/parameter/ParameterDimming.java
@@ -40,9 +40,9 @@ public interface ParameterDimming extends IObj
 
 	public static final String HREF = "/parameters/dimming";
 
-	public static final int MAX_VALUE = 100;
+	public static final long MAX_VALUE = 100;
 
-	public static final int MIN_VALUE = 0;
+	public static final long MIN_VALUE = 0;
 
 	public Int value();
 }
